=General *Added full support for the V5 item packets. -These packets allows the use of the ShadoW Equip slots. -These slots functions like the other slots, allowing the use of -equips with slots, refines, scripts, etc. A 2013-12-23Ragexe -client or higher is required to use this.
*clif_item_sub_v5
*clif_item_flag_v5
*clif_inventorylist_v5
*clif_equiplist_v5
*clif_storagelist_v5
*clif_cartlist_v5
*clif_viewequip_ack_v5
-Added support for these functions.
-These functions are used for the V5 item packets since the structure
-has changed in a way that would make it difficult to merge into the
-older item packet functions for some of the packets.

git-svn-id: https://svn.code.sf.net/p/v1-3ceam/code/trunk@762 cad27aaa-dce3-4a30-a00a-e4fd67c11881

+++ rewrite/src/common/mmo.h
@@ -234,7 +234,7 @@ struct item {
 	int id;
 	short nameid;
 	short amount;
-	unsigned short equip; // location(s) where item is equipped (using enum equip_pos for bitmasking)
+	unsigned int equip; // location(s) where item is equipped (using enum equip_pos for bitmasking)
 	char identify;
 	char refine;
 	char attribute;
